+++ exporter/datadogexporter/config/config.go
@@ -268,6 +268,9 @@ type Config struct {
 	warnings []error
 }
 
+// OnceMetadata gets a sync.Once instance used for initializing the host metadata.
+// Deprecated: [v0.48.0] do not use, will be removed on v0.49.0.
+// TODO (#8373): Remove this method.
 func (c *Config) OnceMetadata() *sync.Once {
 	return &c.onceMetadata
 }