+++ exporter/newrelicexporter/transformer.go
@@ -413,7 +413,7 @@ func (t *transformer) MetricAttributes(baseAttributes map[string]interface{}, at
 	for k, v := range baseAttributes {
 		rawMap[k] = v
 	}
-	attrMap.Range(func(k string, v pdata.AttributeValue) bool {
+	attrMap.Range(func(k string, v pdata.Value) bool {
 		// Only include attribute if not an override attribute
 		if _, isOverrideKey := t.OverrideAttributes[k]; !isOverrideKey {
 			rawMap[k] = v.AsString()
@@ -425,7 +425,7 @@ func (t *transformer) MetricAttributes(baseAttributes map[string]interface{}, at
 }
 
 func (t *transformer) TrackAttributes(location attributeLocation, attributeMap pdata.AttributeMap) {
-	attributeMap.Range(func(_ string, v pdata.AttributeValue) bool {
+	attributeMap.Range(func(_ string, v pdata.Value) bool {
 		statsKey := attributeStatsKey{location: location, attributeType: v.Type()}
 		t.details.attributeMetadataCount[statsKey]++
 		return true