+++ exporter/signalfxexporter/internal/translation/logdata_to_signalfxv2.go
@@ -55,13 +55,13 @@ func convertLogRecord(lr pdata.LogRecord, resourceAttrs pdata.AttributeMap, logg
 
 	var event sfxpb.Event
 
-	if categoryVal.Type() == pdata.AttributeValueTypeInt {
+	if categoryVal.Type() == pdata.ValueTypeInt {
 		asCat := sfxpb.EventCategory(categoryVal.IntVal())
 		event.Category = &asCat
 	}
 
-	if mapVal, ok := attrs.Get(splunk.SFxEventPropertiesKey); ok && mapVal.Type() == pdata.AttributeValueTypeMap {
-		mapVal.MapVal().Range(func(k string, v pdata.AttributeValue) bool {
+	if mapVal, ok := attrs.Get(splunk.SFxEventPropertiesKey); ok && mapVal.Type() == pdata.ValueTypeMap {
+		mapVal.MapVal().Range(func(k string, v pdata.Value) bool {
 			val, err := attributeValToPropertyVal(v)
 			if err != nil {
 				logger.Debug("Failed to convert log record property value to SignalFx property value", zap.Error(err), zap.String("key", k))
@@ -79,7 +79,7 @@ func convertLogRecord(lr pdata.LogRecord, resourceAttrs pdata.AttributeMap, logg
 	// keep a record of Resource attributes to add as dimensions
 	// so as not to modify LogRecord attributes
 	resourceAttrsForDimensions := pdata.NewAttributeMap()
-	resourceAttrs.Range(func(k string, v pdata.AttributeValue) bool {
+	resourceAttrs.Range(func(k string, v pdata.Value) bool {
 		// LogRecord attribute takes priority
 		if _, ok := attrs.Get(k); !ok {
 			resourceAttrsForDimensions.Insert(k, v)
@@ -87,7 +87,7 @@ func convertLogRecord(lr pdata.LogRecord, resourceAttrs pdata.AttributeMap, logg
 		return true
 	})
 
-	addDimension := func(k string, v pdata.AttributeValue) bool {
+	addDimension := func(k string, v pdata.Value) bool {
 		// Skip internal attributes
 		switch k {
 		case splunk.SFxEventCategoryKey:
@@ -95,13 +95,13 @@ func convertLogRecord(lr pdata.LogRecord, resourceAttrs pdata.AttributeMap, logg
 		case splunk.SFxEventPropertiesKey:
 			return true
 		case splunk.SFxEventType:
-			if v.Type() == pdata.AttributeValueTypeString {
+			if v.Type() == pdata.ValueTypeString {
 				event.EventType = v.StringVal()
 			}
 			return true
 		}
 
-		if v.Type() != pdata.AttributeValueTypeString {
+		if v.Type() != pdata.ValueTypeString {
 			logger.Debug("Failed to convert log record or resource attribute value to SignalFx property value, key is not a string", zap.String("key", k))
 			return true
 		}
@@ -123,19 +123,19 @@ func convertLogRecord(lr pdata.LogRecord, resourceAttrs pdata.AttributeMap, logg
 	return &event, true
 }
 
-func attributeValToPropertyVal(v pdata.AttributeValue) (*sfxpb.PropertyValue, error) {
+func attributeValToPropertyVal(v pdata.Value) (*sfxpb.PropertyValue, error) {
 	var val sfxpb.PropertyValue
 	switch v.Type() {
-	case pdata.AttributeValueTypeInt:
+	case pdata.ValueTypeInt:
 		asInt := v.IntVal()
 		val.IntValue = &asInt
-	case pdata.AttributeValueTypeBool:
+	case pdata.ValueTypeBool:
 		asBool := v.BoolVal()
 		val.BoolValue = &asBool
-	case pdata.AttributeValueTypeDouble:
+	case pdata.ValueTypeDouble:
 		asDouble := v.DoubleVal()
 		val.DoubleValue = &asDouble
-	case pdata.AttributeValueTypeString:
+	case pdata.ValueTypeString:
 		asString := v.StringVal()
 		val.StrValue = &asString
 	default: