+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 ## Unreleased
+- `sigv4authextension`: Enable component (#8518)
+
+### 💡 Enhancements 💡
+
+- `k8seventsreceiver`: Add Api_version and resource_version (#8539)
+
+### 🛑 Breaking changes 🛑
+
+- `windowsperfcountersreceiver`: Added metrics configuration (#8376)
+
+### 🚩 Deprecations 🚩
+
+- `datadogexporter`: Deprecate `OnlyMetadata` method from `Config` struct (#8359)
 
 - `resourcedetectionprocessor`: Add attribute allowlist (#8547)
 
@@ -8,6 +21,7 @@
 
 ### 💡 Enhancements 💡
 
+- `googlecloudexporter`: Add Validate method in config (#8559)
 - `attributesprocessor`: Add convert action (#7930)
 - `attributesprocessor`: Add metric support (#8111)
 - `prometheusremotewriteexporter`: Write-Ahead Log support enabled (#7304)
@@ -28,6 +42,7 @@
 - `coralogixexporter` Allow exporter timeout to be configured (#7957)
 - `prometheusremotewriteexporter` support adding trace id and span id attached to exemplars (#8380)
 - `influxdbexporter`: accept histogram metric missing infinity bucket. (#8462)
+- `skywalkingreceiver`: Added implementation of Skywalking receiver. (#8549)
 
 ### 🛑 Breaking changes 🛑